resizing and Redraw

+++ Assets/Scripts/FileThumbnail.cs
@@ -16,21 +16,30 @@ using Debug = UnityEngine.Debug;
 public static class FileThumbnail
 {
     public static readonly string CacheThumbnailFile =
-        Path.Combine(Application.persistentDataPath, "cache.png").Replace('\\', '/');
+        Path.Combine(Application.persistentDataPath, "cache.png")
+            .Replace('/', Path.DirectorySeparatorChar)
+            .Replace('\\', Path.DirectorySeparatorChar);
 
     public static readonly string ThumbnailSolution =
-        Path.Combine(Application.dataPath.Replace("Assets", "Thumbnails"), "Thumbnails.csproj").Replace('\\', '/');
+        Path.Combine(Application.dataPath.Replace("Assets", "Thumbnails"), "Thumbnails.csproj")
+            .Replace('/', Path.DirectorySeparatorChar)
+            .Replace('\\', Path.DirectorySeparatorChar);
 #if UNITY_EDITOR
     public static readonly string BuildArtifacts =
         Path.Combine(Application.dataPath.Replace("Assets", "Build"), $"{Application.productName}_Data/Thumbnails")
-            .Replace('\\', '/');
+            .Replace('/', Path.DirectorySeparatorChar)
+            .Replace('\\', Path.DirectorySeparatorChar);
 #else
     public static readonly string BuildArtifacts =
-        Path.Combine(Application.dataPath.Replace("Assets", "Build"), "Thumbnails").Replace('\\', '/');
+        Path.Combine(Application.dataPath.Replace("Assets", "Build"), "Thumbnails")
+            .Replace('/', Path.DirectorySeparatorChar)
+            .Replace('\\', Path.DirectorySeparatorChar);
 #endif
 
     public static readonly string ThumbnailsExecutable =
-        Path.Combine(BuildArtifacts, "bin/Thumbnails/debug/Thumbnails.exe").Replace('\\', '/');
+        Path.Combine(BuildArtifacts, "bin/Thumbnails/debug/Thumbnails.exe")
+            .Replace('/', Path.DirectorySeparatorChar)
+            .Replace('\\', Path.DirectorySeparatorChar);
 
 #if UNITY_EDITOR
     [MenuItem("File/Build Thumbnails", false, 3)]
@@ -46,7 +55,8 @@ public static class FileThumbnail
             await UniTask.WaitWhile(() => !process.HasExited);
         }
 
-        Directory.Delete(Path.Combine(BuildArtifacts, "obj"), true);
+        if (Directory.Exists(Path.Combine(BuildArtifacts, "obj")))
+            Directory.Delete(Path.Combine(BuildArtifacts, "obj"), true);
     }
 #endif
 
@@ -99,6 +109,18 @@ namespace CustomHelper
                 bitmap.Save(ms, ImageFormat.Png);
                 texture.LoadImage(ms.ToArray());
             }
+
+            if (bitmap.GetPixel(96, 96).A == 0 && bitmap.GetPixel(160, 160).A == 0)
+            {
+                texture.Apply();
+                var newTexture = new Texture2D(64, 64, TextureFormat.ARGB32, false);
+                newTexture.SetPixels(0, 0, 64, 64, texture.GetPixels(96, 96, 64, 64));
+                newTexture.Apply();
+                texture = newTexture;
+            }
+
+            if (texture.height == 64)
+                texture.filterMode = FilterMode.Point;
 
             texture.Apply();
             return texture;
